requirements.txt
pynput
python-dotenv
schedule
memory-profiler
GitPython